misc: move UI types/interfaces to `src/@types/ui-types.ts` (#7365)
