+++ src/ui/handlers/login-form-ui-handler.ts
@@ -1,8 +1,8 @@
 import { pokerogueApi } from "#api/api";
 import { globalScene } from "#app/global-scene";
 import { UiMode } from "#enums/ui-mode";
+import type { ModalConfig } from "#types/ui-types";
 import type { InputFieldConfig } from "#ui/form-modal-ui-handler";
-import type { ModalConfig } from "#ui/modal-ui-handler";
 import { OAuthProvidersUiHandler } from "#ui/oauth-providers-ui-handler";
 import i18next from "i18next";
