+++ src/ui/handlers/pokedex-scan-ui-handler.ts
@@ -2,11 +2,10 @@ import { speciesDataRegistry } from "#app/global-species-data-registry";
 import { allAbilities, allMoves } from "#data/data-lists";
 import { UiMode } from "#enums/ui-mode";
 import type { PlayerPokemon } from "#field/pokemon";
-import type { OptionSelectItem } from "#ui/base-option-select-ui-handler";
+import type { ModalConfig, OptionSelectItem } from "#types/ui-types";
 import { FilterTextRow } from "#ui/filter-text";
 import type { InputFieldConfig } from "#ui/form-modal-ui-handler";
 import { FormModalUiHandler } from "#ui/form-modal-ui-handler";
-import type { ModalConfig } from "#ui/modal-ui-handler";
 import i18next from "i18next";
 
 export class PokedexScanUiHandler extends FormModalUiHandler {
