+++ src/ui/handlers/registration-form-ui-handler.ts
@@ -3,8 +3,8 @@ import { globalScene } from "#app/global-scene";
 import { TextStyle } from "#enums/text-style";
 import { UiMode } from "#enums/ui-mode";
 import type { LoginPhase } from "#phases/login-phase";
+import type { ModalConfig } from "#types/ui-types";
 import type { InputFieldConfig } from "#ui/form-modal-ui-handler";
-import type { ModalConfig } from "#ui/modal-ui-handler";
 import { addTextObject } from "#ui/text";
 import { fixedInt } from "#utils/common";
 import i18next from "i18next";
