+++ src/ui/handlers/rename-form-ui-handler.ts
@@ -1,7 +1,7 @@
 import type { PlayerPokemon } from "#field/pokemon";
+import type { ModalConfig } from "#types/ui-types";
 import type { InputFieldConfig } from "#ui/form-modal-ui-handler";
 import { FormModalUiHandler } from "#ui/form-modal-ui-handler";
-import type { ModalConfig } from "#ui/modal-ui-handler";
 import i18next from "i18next";
 
 export class RenameFormUiHandler extends FormModalUiHandler {
