+++ src/ui/handlers/target-select-ui-handler.ts
@@ -7,10 +7,9 @@ import { UiMode } from "#enums/ui-mode";
 import type { Pokemon } from "#field/pokemon";
 import type { ModifierBar } from "#modifiers/modifier";
 import { getMoveTargets } from "#moves/move-utils";
+import type { TargetSelectCallback } from "#types/ui-types";
 import { UiHandler } from "#ui/ui-handler";
 import { fixedInt } from "#utils/common";
-
-export type TargetSelectCallback = (targets: BattlerIndex[]) => void;
 
 export class TargetSelectUiHandler extends UiHandler {
   private fieldIndex: number;
