+++ src/ui/handlers/test-dialogue-ui-handler.ts
@@ -1,9 +1,8 @@
 import { UiMode } from "#enums/ui-mode";
 import type { PlayerPokemon } from "#field/pokemon";
-import type { OptionSelectItem } from "#ui/base-option-select-ui-handler";
+import type { ModalConfig, OptionSelectItem } from "#types/ui-types";
 import type { InputFieldConfig } from "#ui/form-modal-ui-handler";
 import { FormModalUiHandler } from "#ui/form-modal-ui-handler";
-import type { ModalConfig } from "#ui/modal-ui-handler";
 import i18next from "i18next";
 
 /** Handler for the i18n testing menu used by the localization team. */
