+++ src/ui/settings/base-settings-ui-handler.ts
@@ -5,10 +5,10 @@ import { UiMode } from "#enums/ui-mode";
 import type { SettingType } from "#system/settings";
 import { Setting, SettingKeys } from "#system/settings";
 import type { MappingSettingName } from "#types/configs/inputs";
+import type { InputsIcons } from "#types/ui-types";
 import { TabMenu } from "#ui/containers/tab-menu";
 import { MessageUiHandler } from "#ui/message-ui-handler";
 import { ScrollBar } from "#ui/scroll-bar";
-import type { InputsIcons } from "#ui/settings/base-control-settings-ui-handler";
 import { addTextObject, getTextColor } from "#ui/text";
 import type { TitleUiHandler } from "#ui/title-ui-handler";
 import { addWindow } from "#ui/ui-theme";
