+++ src/ui/settings/special-icons.ts
@@ -2,4 +2,5 @@ export const specialIcons = {
   BUTTON_HOME: "HOME.png",
   BUTTON_DELETE: "DEL.png",
 };
+
 export const specialIconKeys = Object.keys<string>(specialIcons);
